+++ misskey-api/src/endpoint.rs
@@ -93,4 +93,16 @@ pub mod stats;
 
 #[cfg(feature = "13-0-0")]
 #[cfg_attr(docsrs, doc(cfg(feature = "13-0-0")))]
+pub mod emojis;
+
+#[cfg(feature = "13-0-0")]
+#[cfg_attr(docsrs, doc(cfg(feature = "13-0-0")))]
 pub mod flash;
+
+#[cfg(feature = "13-0-0")]
+#[cfg_attr(docsrs, doc(cfg(feature = "13-0-0")))]
+pub mod invite;
+
+#[cfg(feature = "13-0-0")]
+#[cfg_attr(docsrs, doc(cfg(feature = "13-0-0")))]
+pub mod retention;