+++ misskey-api/src/endpoint/notes/state.rs
@@ -12,6 +12,8 @@ pub struct Request {
 #[serde(rename_all = "camelCase")]
 pub struct Response {
     pub is_favorited: bool,
+    #[cfg(not(feature = "13-0-0"))]
+    #[cfg_attr(docsrs, doc(cfg(not(feature = "13-0-0"))))]
     pub is_watching: bool,
     #[cfg(feature = "12-95-0")]
     #[cfg_attr(docsrs, doc(cfg(feature = "12-95-0")))]