+++ misskey-api/src/model/emoji.rs
@@ -1,6 +1,7 @@
 use crate::model::id::Id;
 
 use serde::{Deserialize, Serialize};
+#[cfg(not(feature = "13-0-0"))]
 use url::Url;
 
 #[derive(Serialize, Deserialize, Debug, Clone)]
@@ -8,6 +9,8 @@ use url::Url;
 pub struct Emoji {
     pub id: Id<Emoji>,
     pub name: String,
+    #[cfg(not(feature = "13-0-0"))]
+    #[cfg_attr(docsrs, doc(cfg(not(feature = "13-0-0"))))]
     pub url: Url,
     pub host: Option<String>,
     pub category: Option<String>,
@@ -15,3 +18,13 @@ pub struct Emoji {
 }
 
 impl_entity!(Emoji);
+
+#[cfg(feature = "13-0-0")]
+#[cfg_attr(docsrs, doc(cfg(feature = "13-0-0")))]
+#[derive(Deserialize, Debug, Clone)]
+#[serde(rename_all = "camelCase")]
+pub struct EmojiSimple {
+    pub name: String,
+    pub aliases: Vec<String>,
+    pub category: Option<String>,
+}